refactor: LotteryResult의 BuyerResult 생성 로직 BuyerResult로 이동
